+++ Paramount.Betterclassifieds.Presentation/Scripts/app/events/addGuest.model.js
@@ -1,11 +1,13 @@
 ﻿(function ($, $p, ko) {
-    
+
     function AddGuest() {
         this.id = ko.observable();
         this.eventId = ko.observable();
+        this.isSeatedEvent = ko.observable();
         this.guestFullName = ko.observable();
         this.guestEmail = ko.observable();
         this.selectedTicket = ko.observable();
+        this.seatNumber = ko.observable();
         this.sendEmailToGuest = ko.observable(true);
         this.tickets = ko.observableArray();
         this.ticketFields = ko.observableArray();
@@ -15,13 +17,47 @@
         this.isPublic = ko.observable();
         this.hasGroups = ko.observable(false);
         this.eventGroups = ko.observableArray();
+        this.seats = ko.observableArray();
+        this.guestAddWarning = ko.observable('');
         this.validator = ko.validatedObservable({
             guestFullName: this.guestFullName.extend({ required: true }),
             guestEmail: this.guestEmail.extend({ required: true, email: true })
         });
+        this.isLoading = ko.observable(true);
+        this.cachedEventSeating = [];
+
+        var me = this;
+        me.seatSelected = function (seatNumber) {
+            me.guestAddWarning('');
+            var seat = _.find(me.cachedEventSeating, function (s) {
+                return s.seatNumber === seatNumber;
+            });
+            if (seat.available === false) {
+                me.guestAddWarning('The selected seat is not available.');
+                me.seatNumber(null);
+                me.selectedTicket(null);
+            } else {
+                var eventTicketId = seat.eventTicketId;
+                var eventTicket = _.find(me.tickets(), function (t) {
+                    return t.eventTicketId() === eventTicketId;
+                });
+                me.selectedTicket(eventTicket);
+                me.ticketChanged();
+            }
+        }
+
+        me.getSelectedTicketName = ko.computed(function () {
+            return me.selectedTicket() && me.selectedTicket().ticketName();
+        });
+
+        me.getSelectedTicketRemainingCount = ko.computed(function () {
+            return me.selectedTicket() && me.selectedTicket().remainingQuantity();
+        });
     }
 
     AddGuest.prototype.bindData = function (data) {
+        var service = new $p.EventService();
+
         var me = this;
         me.id(data.id);
         me.eventId(data.eventId);
@@ -30,6 +66,7 @@
         me.displayGuests(data.displayGuests);
         me.isPublic(data.isPublic);
         me.hasGroups(data.eventGroups && data.eventGroups.length > 0);
+        me.isSeatedEvent(data.isSeatedEvent);
 
         _.each(data.ticketFields, function (tf) {
             me.ticketFields.push(new $p.models.DynamicFieldValue(tf));
@@ -42,25 +79,60 @@
         _.each(data.eventGroups, function (gr) {
             me.eventGroups.push(new $p.models.EventGroup(gr));
         });
+
+        if (data.isSeatedEvent === true) {
+            service.getEventSeating(data.eventId).then(function (response) {
+                if (!response.rows) {
+                    toastr.error('Unable to find any seats. Please ensure the seating has been setup first.');
+                }
+                _.each(response.rows, function (r) {
+                    _.each(r.seats, function (s) {
+                        me.seats.push(s.seatNumber);
+                        me.cachedEventSeating.push(s);
+                    });
+                });
+            });
+        } else {
+            me.isLoading(false);
+        }
     }
 
     AddGuest.prototype.submitGuest = function (element, event) {
         var me = this;
-        if ( !$p.checkValidity(me) || !$p.checkValidity(me.ticketFields()) ) {
+        if (!$p.checkValidity(me) || !$p.checkValidity(me.ticketFields())) {
+            return;
+        }
+
+        if (me.isSeatedEvent() === true && me.seatNumber() === null || me.seatNumber() === undefined) {
+            me.guestAddWarning('Please ensure seat is selected');
             return;
         }
 
         var $btn = $(event.target);
         $btn.button('loading');
 
-        var objToSend = ko.toJS(me);
-        delete objToSend["tickets"]; // Remove unecessary collection that is bound to a simple dropdown
+        var objToSend = this.toModel(me);
 
         var service = new $p.AdDesignService(me.id());
         service.addGuest(objToSend).then(function () {
             $btn.button('reset');
             me.saved(true);
         });
+    }
+
+    AddGuest.prototype.toModel = function (vm) {
+        // maps to AddEventGuestViewModel
+        return {
+            id: vm.id(),
+            eventId: vm.eventId(),
+            guestFullName: vm.guestFullName(),
+            guestEmail: vm.guestEmail(),
+            isPublic: vm.isPublic(),
+            seatNumber: vm.seatNumber(),
+            sendEmailToGuest: vm.sendEmailToGuest(),
+            selectedTicket: ko.toJS(vm.selectedTicket()),
+            selectedGroup: ko.toJS(vm.selectedGroup)
+        }
     }
 
     AddGuest.prototype.ticketChanged = function () {
@@ -74,7 +146,7 @@
         });
 
         me.eventGroups.removeAll();
-        service.getGroupsForTicket(me.eventId(), me.selectedTicket().eventTicketId()).then(function(resp) {
+        service.getGroupsForTicket(me.eventId(), me.selectedTicket().eventTicketId()).then(function (resp) {
             _.each(resp, function (gr) {
                 me.eventGroups.push(new $p.models.EventGroup(gr));
             });
@@ -84,9 +156,15 @@
     AddGuest.prototype.addAnother = function (element, event) {
         this.guestFullName(null);
         this.guestEmail(null);
+        this.selectedTicket(null);
+        this.selectedGroup(null);
+        this.seatNumber(null);
+
         _.each(this.ticketFields(), function (tf) {
             tf.fieldValue(null);
         });
+        // Reset all warnings
+        this.guestAddWarning('');
         this.saved(false);
         $('html, body').animate({ scrollTop: $('.add-guest-view').offset().top }, 1000);
     }